backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.35.0
sqlalchemy>=2.0.35
alembic>=1.13.3
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-dotenv>=1.1.0
python-multipart>=0.0.12
httpx>=0.28.1
pytest>=8.3.3
pytest-asyncio>=0.24.0
